Fixed bug which sometimes broke challenge generation on PHP 8.1
Sometimes on PHP 8.1, the challenge image would not generate and instead show a broken image. This bug was caused by a deprecation warning, which is thrown when floats are implicitly converted to integers (deprecated: Implicit conversion from float to int loses precision). It only occurred when generating challenges with uneven icon amounts.

+++ src/captcha.class.php
@@ -447,9 +447,9 @@ class IconCaptcha
         // Image pixel information.
         $iconCount = count(self::$session->icons);
         $iconSize = self::CAPTCHA_ICON_SIZES[$iconCount];
-        $iconOffset = ((self::CAPTCHA_IMAGE_SIZE / $iconCount) - 30) / 2;
-        $iconOffsetAdd = (self::CAPTCHA_IMAGE_SIZE / $iconCount) - $iconSize;
-        $iconLineSize = self::CAPTCHA_IMAGE_SIZE / $iconCount;
+        $iconOffset = (int)(((self::CAPTCHA_IMAGE_SIZE / $iconCount) - 30) / 2);
+        $iconOffsetAdd = (int)((self::CAPTCHA_IMAGE_SIZE / $iconCount) - $iconSize);
+        $iconLineSize = (int)(self::CAPTCHA_IMAGE_SIZE / $iconCount);
 
         // Options.
         $rotateEnabled = self::$options['image']['rotate'];
